Refactor text2vec-wcs -> text2vec-weaviate
