+++ modules/text2vec-weaviate/clients/wcs_embed_test.go
@@ -95,7 +95,7 @@ func TestClient(t *testing.T) {
 		assert.Equal(t, err.Error(), "WCS embed API error: 500 ")
 	})
 
-	t.Run("when WCS key is passed using X-WCS-Api-Key header", func(t *testing.T) {
+	t.Run("when Weaviate Inference key is passed using X-Weaviate-Inference-Key header", func(t *testing.T) {
 		server := httptest.NewServer(&fakeHandler{t: t})
 		defer server.Close()
 		c := &vectorizer{
@@ -108,7 +108,7 @@ func TestClient(t *testing.T) {
 			logger: nullLogger(),
 		}
 		ctxWithValue := context.WithValue(context.Background(),
-			"X-WCS-Api-Key", []string{"some-key"})
+			"X-Weaviate-Inference-Key", []string{"some-key"})
 
 		expected := &modulecomponents.VectorizationResult{
 			Text:       []string{"This is my text"},
@@ -121,7 +121,7 @@ func TestClient(t *testing.T) {
 		assert.Equal(t, expected, res)
 	})
 
-	t.Run("when WCS key is empty", func(t *testing.T) {
+	t.Run("when Weaviate Inference key is empty", func(t *testing.T) {
 		server := httptest.NewServer(&fakeHandler{t: t})
 		defer server.Close()
 		c := &vectorizer{
@@ -139,12 +139,12 @@ func TestClient(t *testing.T) {
 		_, _, err := c.Vectorize(ctx, []string{"This is my text"}, fakeClassConfig{classConfig: map[string]interface{}{"baseURL": server.URL}})
 
 		require.NotNil(t, err)
-		assert.Equal(t, err.Error(), "WCS embed API key: no api key found "+
-			"neither in request header: X-WCS-Api-Key "+
-			"nor in environment variable under WCS_APIKEY")
+		assert.Equal(t, err.Error(), "Weaviate Inference API key: no api key found "+
+			"neither in request header: X-Weaviate-Inference-Key "+
+			"nor in environment variable under WEAVIATE_INFERENCE_KEY")
 	})
 
-	t.Run("when X-WCS-Api-Key header is passed but empty", func(t *testing.T) {
+	t.Run("when X-Weaviate-Inference-Key header is passed but empty", func(t *testing.T) {
 		server := httptest.NewServer(&fakeHandler{t: t})
 		defer server.Close()
 		c := &vectorizer{
@@ -157,14 +157,14 @@ func TestClient(t *testing.T) {
 			logger: nullLogger(),
 		}
 		ctxWithValue := context.WithValue(context.Background(),
-			"X-WCS-Api-Key", []string{""})
+			"X-Weaviate-Inference-Key", []string{""})
 
 		_, _, err := c.Vectorize(ctxWithValue, []string{"This is my text"}, fakeClassConfig{classConfig: map[string]interface{}{}})
 
 		require.NotNil(t, err)
-		assert.Equal(t, err.Error(), "WCS embed API key: no api key found "+
-			"neither in request header: X-WCS-Api-Key "+
-			"nor in environment variable under WCS_APIKEY")
+		assert.Equal(t, err.Error(), "Weaviate Inference API key: no api key found "+
+			"neither in request header: X-Weaviate-Inference-Key "+
+			"nor in environment variable under WEAVIATE_INFERENCE_KEY")
 	})
 
 	t.Run("when X-WCS-BaseURL header is passed", func(t *testing.T) {
