+++ modules/text2vec-weaviate/ent/class_settings.go
