+++ modules/text2vec-weaviate/ent/vectorization_config.go
