+++ modules/text2vec-weaviate/vectorizer/batch_test.go
