+++ modules/text2vec-weaviate/vectorizer/fakes_for_test.go
